requirements.txt
argostranslate>=1.9.0
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0
markdown-it-py>=3.0.0

